// framework/magento/schema.d.ts
export type CategoryTree = {
  available_sort_by: Array<Scalars["String"]>;
  breadcrumbs: Array<Breadcrumb>;
  canonical_url: Scalars["String"];
  children: Array<CategoryTree>;
  children_count: Scalars["String"];
  cms_block: CmsBlock;
  custom_layout_update_file: Scalars["String"];
  default_sort_by: Scalars["String"];
  description: Scalars["String"];
  display_mode: Scalars["String"];
  filter_price_range: Scalars["Float"];
  image: Scalars["String"];
  include_in_menu: Scalars["Int"];
  is_anchor: Scalars["Int"];
  landing_page: Scalars["Int"];
  level: Scalars["Int"];
  meta_description: Scalars["String"];
  meta_keywords: Scalars["String"];
  meta_title: Scalars["String"];
  name: Scalars["String"];
  path: Scalars["String"];
  path_in_store: Scalars["String"];
  position: Scalars["Int"];
  product_count: Scalars["Int"];
  products: CategoryProducts;
  relative_url: Scalars["String"];
  titlecat: Scalars["String"];
  type: UrlRewriteEntityTypeEnum;
  uid: Scalars["String"];
  url_key: Scalars["String"];
  url_path: Scalars["String"];
  url_suffix: Scalars["String"];
};

export type CmsBlock = {
  content: Scalars["String"];
  identifier: Scalars["String"];
  title: Scalars["String"];
};

export enum UrlRewriteEntityTypeEnum {
  CMS_PAGE = "CMS_PAGE",
  PRODUCT = "PRODUCT",
  CATEGORY = "CATEGORY",
}

export type ProductAttributeSortInput = {
  maxguests: SortEnum;
  name: SortEnum;
  position: SortEnum;
  price: SortEnum;
  relevance: SortEnum;
};

export enum SortEnum {
  ASC = "ASC",
  DESC = "DESC",
}

export type CategoryProducts = {
  items: Array<ProductItem>;
  page_info: SearchResultPageInfo;
  total_count: Scalars["Int"];
};

export type MultiSelectAttributes<T> = {
  MultiSelectAttributes: MultiSelectAttributes;
  [key: T]: T;
};

export type ComplexTextValue = {
  html: Scalars["String"];
};

export type ProductImage = {
  disabled: Scalars["Boolean"];
  label: Scalars["String"];
  position: Scalars["Int"];
  url: Scalars["String"];
};

//ProductLinksInterface
export type ProductLinks = {
  link_type: Scalars["String"];
  linked_product_sku: Scalars["String"];
  linked_product_type: Scalars["String"];
  position: Scalars["Int"];
  sku: Scalars["String"];
};

export type ProductDiscount = {
  amount_off: Scalars["Float"];
  percent_off: Scalars["Float"];
};

export type Money = {
  currency: CurrenyEnum;
  value: Scalars["Float"];
};

export type TierPrice = {
  discount: ProductDiscount;
  final_price: Money;
  quantity: Scalars["Float"];
};

export type PriceRange = {
  maximum_price: ProductPrice;
  minimum_price: ProductPrice!;
};

export type ProductPrice = {
  discount: ProductDiscount;
  final_price: Money;
  fixed_product_taxes: Array<FixedProductTax>;
  regular_price: Money;
};

export type FixedProductTax = {
  amount: Money;
  label: Scalars["String"];
};

export type HttpQueryParameter = {
  name: Scalars["String"];
  value: Scalars["String"];
};

export type UrlRewrite = {
  parameters: Array<HttpQueryParameter>;
  url: Scalars["String"];
};

export enum ProductStockStatus {
  IN_STOCK = "IN_STOCK",
  OUT_OF_STOCK = "OUT_OF_STOCK",
}

export type ProductItem = {
  additional_information: Scalars["String"];
  builder: Scalars["String"];
  canonical_url: Scalars["String"];
  categories: Array<CategoryTree>;
  country_of_manufacture: Scalars["String"];
  crosssell_products: Array<ProductItem[]>;
  customattributes: MultiSelectAttributes;
  description: ComplexTextValue;
  exterior_designer: Scalars["String"];
  gift_message_available: Scalars["String"];
  image: ProductImage;
  media_gallery: ProductImage[];
  meta_description: Scalars["String"];
  meta_keyword: Scalars["String"];
  meta_title: Scalars["String"];
  name: Scalars["String"];
  new_from_date: Scalars["String"];
  new_to_date: Scalars["String"];
  only_x_left_in_stock: Float;
  options_container: Scalars["String"];
  price_range: PriceRange;
  price_tiers: Array<TierPrice>;
  product_links: Array<ProductLinks>;
  productattribute: Scalars["String"];
  related_products: ProductItem[];
  short_description: ComplexTextValue;
  sku: Scalars["String"];
  small_image: ProductImage;
  special_price: Scalars["Float"];
  special_to_date: Scalars["String"];
  stock_status: ProductStockStatus;
  swatch_image: Scalars["String"];
  thumbnail: ProductImage;
  uid: Scalars["ID"];
  upgrades: Scalars["String"];
  upsell_products: ProductItem[];
  url_key: Scalars["String"];
  url_rewrites: Array<UrlRewrite>;
  url_suffix: Scalars["String"];
  videoid: Scalars["String"];
  year: Scalars["String"];
};

export type AggregationOption = {
  count: Scalars["Int"];
  label: Scalars["String"];
  value: Scalars["String"];
};

export type Aggregations = {
  attribute_code: Scalars["String"];
  count: Scalars["Int"];
  label: Scalars["String"];
  options: AggregationOption;
  position: Scalars["Int"];
};

export type SortFields = {
  default: Scalars["String"];
  options: Omit<AggregationOption, "count">;
};

export type Suggestions = {
  search: Scalars["String"];
};

export type ProductCollections = {
  aggregations: Aggregations;
  page_info: SearchResultPageInfo;
  sort_fields: SortFields;
  suggestions: Suggestions;
  total_count: Scalars["Int"];
  items: Array<ProductItem>;
};
// ...
export type SearchResultPageInfo = {
  current_page: Scalars["Int"];
  page_size: Scalars["Int"];
  total_pages: Scalars["Int"];
};

export type Breadcrumb = {
  category_id: Scalars["Int"];
  category_level: Scalars["Int"];
  category_name: Scalars["String"];
  category_uid: Scalars["String"];
  category_url_key: Scalars["String"];
  category_url_path: Scalars["String"];
};
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** An ISO-8601 encoded UTC date time string. Example value: `"2019-07-03T20:47:55Z"`. */
  DateTime: any;
  /** A signed decimal number, which supports arbitrary precision and is serialized as a string. Example value: `"29.99"`. */
  Decimal: any;
  /** A string containing HTML code. Example value: `"<p>Grey cotton knit sweater.</p>"`. */
  HTML: any;
  /** A monetary value string. Example value: `"100.57"`. */
  Money: any;
  /**
   * An RFC 3986 and RFC 3987 compliant URI string.
   *
   * Example value: `"https://johns-apparel.myshopify.com"`.
   *
   */
  URL: any;
};